fix(Settings): Avoid conversion error in case deviceID is -1.
Probably does not actually fix anything, just makes debugging easier.

+++ src/ui/settings/mediasettings.h
@@ -69,7 +69,7 @@ private:
   bool checkAudioSettings();
 
 
-  uint16_t currentDevice_;
+  int currentDevice_;
 
   Ui::CustomSettings *customUI_;
